# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/course": {
            "get": {
                "description": "list the course",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course status, such as over, preparing, in-progress",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "course type, such as ai, mindspore, foundation",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "just list courses of player, if it is set",
                        "name": "mine",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/course/:id/asg/:asgid": {
            "get": {
                "description": "Get assignment",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "GetAssignment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "asg id",
                        "name": "asgid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/course/reginfo": {
            "get": {
                "description": "get register info",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "GetRegisterInfo",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/course/{id}": {
            "get": {
                "description": "get course infomation",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/course/{id}/asg/list": {
            "get": {
                "description": "list assignments",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "ListAssignments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "assignments status, such as finish",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/course/{id}/asg/result": {
            "get": {
                "description": "get submissions",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "GetSubmissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/course/{id}/cert": {
            "get": {
                "description": "get certification",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "GetCertification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/course/{id}/player": {
            "post": {
                "description": "apply the course",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "Apply",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body of applying",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.StudentApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/course/{id}/realted_project": {
            "put": {
                "description": "add related project",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "AddCourseRelatedProject",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "project info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.AddCourseRelatedProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/course/{id}/record": {
            "put": {
                "description": "add play record",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "AddPlayRecord",
                "parameters": [
                    {
                        "type": "string",
                        "description": "course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "record info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.PlayRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.AddCourseRelatedProjectRequest": {
            "type": "object",
            "properties": {
                "owner": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                }
            }
        },
...
        "controller.PlayRecordRequest": {
            "type": "object",
            "properties": {
                "finish_count": {
                    "type": "integer"
                },
                "lesson_id": {
                    "type": "string"
                },
                "play_count": {
                    "type": "integer"
                },
                "point_id": {
                    "type": "string"
                },
                "section_id": {
                    "type": "string"
                }
            }
        },
...
        "controller.StudentApplyRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "detail": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "email": {
                    "type": "string"
                },
                "identity": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                }
            }
        },
...
    }
}